refactor(perps): unify market headers and slide-up market list (#34282)
## **Description**

<!--
Write a short description of the changes included in this pull request,
also include relevant motivation and context. Have in mind the following
questions:
1. What is the reason for the change?
2. What is the improvement/solution?
-->

Lite and Pro market detail screens duplicated header, summary, and
navigation logic. This PR consolidates them into shared
`PerpsMarketHeader` and `PerpsMarketSummary` components, with
Lite-specific actions passed via `endAccessory` and Pro using the
default wallet/watchlist/mode row.

Also fixes Lite layout issues after the migration (top safe-area inset
so the header no longer overlaps the status bar; content-hugging summary
row so the price is not clipped while Pro keeps the fixed 76px summary
height).

When users tap the market identity in the header on Lite or Pro detail
screens, the market list now slides up from the bottom
(`slide_from_bottom`) instead of pushing from the right. Other entry
points (Perps home, search, etc.) keep the default `slide_from_right`.
The tappable identity caret is now a down chevron to match the slide-up
affordance.

## **Changelog**

<!--
If this PR is not End-User-Facing and should not show up in the
CHANGELOG, you can choose to either:
1. Write `CHANGELOG entry: null`
2. Label with `no-changelog`

If this PR is End-User-Facing, please write a short User-Facing
description in the past tense like:
`CHANGELOG entry: Added a new tab for users to see their NFTs`
`CHANGELOG entry: Fixed a bug that was causing some NFTs to flicker`

(This helps the Release Engineer do their job more quickly and
accurately)
-->

CHANGELOG entry: Improved Perps market picker to slide up from the
market detail header

## **Related issues**

<!-- mms-check: type=issue-link required=true -->

Fixes:

## **Manual testing steps**

<!-- mms-check: type=manual-testing required=true -->

```gherkin
Feature: Perps unified market header and market list picker

  Scenario: Lite market detail header and market list from identity
    Given the user is on Perps Lite market detail for any symbol
    Then the header sits below the status bar and the price is fully visible
    When the user taps the market identity (name + down chevron)
    Then the market list slides up from the bottom
    When the user navigates back
    Then the market list dismisses downward and returns to market detail

  Scenario: Pro market detail header and market list from identity
    Given the user is on Perps Pro market detail for any symbol
    Then the summary row height matches the existing Pro layout
    When the user taps the market identity
    Then the market list slides up from the bottom

  Scenario: Market list from Perps home unchanged
    Given the user is on Perps home
    When the user opens the market list via search or "see all"
    Then the market list still slides in from the right
```

## **Screenshots/Recordings**

<!-- mms-check: type=screenshot required=true -->

<!-- If applicable, add screenshots and/or recordings to visualize the
before and after of your change. -->

### **Before**

N/A

### **After**



https://github.com/user-attachments/assets/9e07043f-ebe9-4530-9c2b-5f7f7c3bd513



## **Pre-merge author checklist**

<!--
Every checklist item must be consciously assessed before marking this PR
as
"Ready for review". A checked box means you deliberately considered that
responsibility, not that you literally performed every action listed.

Unchecked boxes are ambiguous: they are not an implicit "N/A" and they
are not
a silent "skip". See `docs/readme/ready-for-review.md` for the full
checklist
semantics.
-->

- [x] I've followed [MetaMask Contributor
Docs](https://github.com/MetaMask/contributor-docs) and [MetaMask Mobile
Coding
Standards](https://github.com/MetaMask/metamask-mobile/blob/main/.github/guidelines/CODING_GUIDELINES.md).
- [x] I've completed the PR template to the best of my ability
- [x] I've included tests if applicable
- [x] I've documented my code using [JSDoc](https://jsdoc.app/) format
if applicable
- [x] I've applied the right labels on the PR (see [labeling
guidelines](https://github.com/MetaMask/metamask-mobile/blob/main/.github/guidelines/LABELING_GUIDELINES.md)).
Not required for external contributors.

#### Performance checks (if applicable)

- [x] I've tested on Android
  - Ideally on a mid-range device; emulator is acceptable
- [x] I've tested with a power user scenario
- Use these [power-user
SRPs](https://consensyssoftware.atlassian.net/wiki/spaces/TL1/pages/edit-v2/401401446401?draftShareId=9d77e1e1-4bdc-4be1-9ebb-ccd916988d93)
to import wallets with many accounts and tokens
- [x] I've instrumented key operations with Sentry traces for production
performance metrics
- See [`trace()`](/app/util/trace.ts) for usage and
[`addToken`](/app/components/Views/AddAsset/components/AddCustomToken/AddCustomToken.tsx#L274)
for an example

For performance guidelines and tooling, see the [Performance
Guide](https://consensyssoftware.atlassian.net/wiki/spaces/TL1/pages/400085549067/Performance+Guide+for+Engineers).

## **Pre-merge reviewer checklist**

<!--
Reviewer checklist items follow the same semantics as the author
checklist: an
unchecked box is ambiguous, a checked box means the reviewer consciously
assessed that responsibility. See `docs/readme/ready-for-review.md`.
-->

- [ ] I've manually tested the PR (e.g. pull and build branch, run the
app, test code being changed).
- [ ] I confirm that this PR addresses all acceptance criteria described
in the ticket it closes and includes the necessary testing evidence such
as recordings and or screenshots.

Made with [Cursor](https://cursor.com)

---------
