+++ app/components/UI/Perps/Views/PerpsMarketListView/PerpsMarketListView.tsx
@@ -55,6 +55,7 @@ import {
   RouteProp,
   useNavigation,
   useFocusEffect,
+  CommonActions,
   StackActions,
 } from '@react-navigation/native';
 import type { AppNavigationProp } from '../../../../../core/NavigationService/types';
@@ -105,6 +106,7 @@ const PerpsMarketListView = ({
   const defaultSortOptionId = route.params?.defaultSortOptionId;
   const defaultSortDirection = route.params?.defaultSortDirection;
   const transactionActiveAbTests = route.params?.transactionActiveAbTests;
+  const replaceOnSelect = route.params?.replaceOnSelect === true;
 
   const isWatchlistEnabled = useSelector(selectPerpsWatchlistEnabledFlag);
   const isRecentlyViewedEnabled = useSelector(
@@ -293,26 +295,54 @@ const PerpsMarketListView = ({
           source_section = PERPS_EVENT_VALUE.SOURCE_SECTION.ALL_MARKETS;
         }
 
+        const detailsParams = {
+          market,
+          source: PERPS_EVENT_VALUE.SOURCE.PERP_MARKETS,
+          source_section,
+          ...(transactionActiveAbTests?.length
+            ? { transactionActiveAbTests }
+            : {}),
+        };
+
+        if (replaceOnSelect) {
+          // Header slide-up picker: dismiss this list and replace the stale
+          // MARKET_DETAILS beneath it so Back does not return to the prior market.
+          navigation.dispatch((state) => {
+            const routes = state.routes.slice(0, -1);
+            if (
+              routes[routes.length - 1]?.name === Routes.PERPS.MARKET_DETAILS
+            ) {
+              routes.pop();
+            }
+            return CommonActions.reset({
+              ...state,
+              index: routes.length,
+              routes: [
+                ...routes,
+                {
+                  name: Routes.PERPS.MARKET_DETAILS,
+                  params: detailsParams,
+                },
+              ],
+            });
+          });
+          return;
+        }
+
         // Use push instead of navigate so that MARKET_LIST is always beneath
         // MARKET_DETAILS in the stack. navigate() can jump to an existing
         // MARKET_DETAILS entry (e.g. one opened from PerpsHome via the watchlist
         // component's ROOT-based navigation), which would skip MARKET_LIST on
         // back and land the user on PERPS_HOME instead.
         navigation.dispatch(
-          StackActions.push(Routes.PERPS.MARKET_DETAILS, {
-            market,
-            source: PERPS_EVENT_VALUE.SOURCE.PERP_MARKETS,
-            source_section,
-            ...(transactionActiveAbTests?.length
-              ? { transactionActiveAbTests }
-              : {}),
-          }),
+          StackActions.push(Routes.PERPS.MARKET_DETAILS, detailsParams),
         );
       }
     },
     [
       onMarketSelect,
       navigation,
+      replaceOnSelect,
       transactionActiveAbTests,
       searchQuery,
       showFavoritesOnly,
