+++ app/components/UI/Perps/components/PerpsMarketHeader/index.ts
@@ -1,1 +1,10 @@
 export { default } from './PerpsMarketHeader';
+export { PERPS_MARKET_HEADER_HEIGHT } from './PerpsMarketHeader';
+export type {
+  PerpsMarketHeaderProps,
+  PerpsMarketHeaderTestIDs,
+} from './PerpsMarketHeader.types';
+export {
+  createLiteMarketHeaderTestIDs,
+  createProMarketHeaderTestIDs,
+} from './perpsMarketHeaderTestIds';
