+++ app/components/UI/Perps/components/PerpsMarketInlineHeader/PerpsMarketInlineHeader.tsx
@@ -14,7 +14,6 @@ import { PERPS_COLLATERAL_SYMBOL } from '../../constants/perpsConfig';
 import { PerpsMarketHeaderSelectorsIDs } from '../../Perps.testIds';
 import LivePriceHeader from '../LivePriceDisplay/LivePriceHeader';
 import PerpsLeverage from '../PerpsLeverage/PerpsLeverage';
-import PerpsMarketIdentity from '../PerpsMarketIdentity';
 import PerpsTokenLogo from '../PerpsTokenLogo';
 
 export interface PerpsMarketInlineHeaderProps {
@@ -27,22 +26,6 @@ export interface PerpsMarketInlineHeaderProps {
   isFavorite?: boolean;
   testID?: string;
   endAccessory?: ReactNode;
-  /**
-   * When true, renders the redesigned market-detail layout instead of the
-   * compact one: title becomes the full asset name, a leverage pill is added
-   * on the first row, the description becomes the static
-   * `[Ticker]-[collateral] perp` pair on the second row, and the live
-   * price/24h change are removed from the header (relocated below by the
-   * parent). When false, the compact layout (pair title + live price) renders.
-   */
-  useDetailLayout?: boolean;
-  /**
-   * Detail layout only. When provided, the market identity (icon + name +
-   * ticker + leverage) becomes a tightly-bounded pressable box that hugs its
-   * content — pressing it fires this callback and shows a button-like pressed
-   * background. The surrounding header row stays non-interactive.
-   */
-  onIdentityPress?: () => void;
 }
 
 export const PerpsMarketInlineHeader = ({
@@ -54,14 +37,9 @@ export const PerpsMarketInlineHeader = ({
   isFavorite = false,
   testID,
   endAccessory,
-  useDetailLayout = false,
-  onIdentityPress,
 }: PerpsMarketInlineHeaderProps) => {
   const displaySymbol = getPerpsDisplaySymbol(market.symbol);
-
-  const displayTitle = useDetailLayout
-    ? market.name || displaySymbol
-    : `${displaySymbol}-${PERPS_COLLATERAL_SYMBOL}`;
+  const displayTitle = `${displaySymbol}-${PERPS_COLLATERAL_SYMBOL}`;
 
   const leverageBadge = useMemo(
     () =>
@@ -70,35 +48,6 @@ export const PerpsMarketInlineHeader = ({
       ) : null,
     [market.maxLeverage],
   );
-
-  const detailIdentity = useMemo(() => {
-    if (!useDetailLayout) {
-      return null;
-    }
-
-    return (
-      <PerpsMarketIdentity
-        symbol={market.symbol}
-        name={market.name}
-        maxLeverage={market.maxLeverage}
-        size={40}
-        gap={3}
-        onPress={onIdentityPress}
-        testIDs={{
-          assetIcon: PerpsMarketHeaderSelectorsIDs.ASSET_ICON,
-          assetName: PerpsMarketHeaderSelectorsIDs.ASSET_NAME,
-          subtitle: PerpsMarketHeaderSelectorsIDs.SUBTITLE,
-          marketListButton: PerpsMarketHeaderSelectorsIDs.MARKET_LIST_BUTTON,
-        }}
-      />
-    );
-  }, [
-    useDetailLayout,
-    market.symbol,
-    market.name,
-    market.maxLeverage,
-    onIdentityPress,
-  ]);
 
   const compactDescription = useMemo(
     () => (
@@ -156,24 +105,16 @@ export const PerpsMarketInlineHeader = ({
       endAccessory={endAccessory}
       endButtonIconProps={endAccessory ? undefined : endButtonIconProps}
       avatar={
-        useDetailLayout ? (
-          detailIdentity
-        ) : (
-          <PerpsTokenLogo
-            symbol={market.symbol}
-            size={40}
-            testID={PerpsMarketHeaderSelectorsIDs.ASSET_ICON}
-          />
-        )
+        <PerpsTokenLogo
+          symbol={market.symbol}
+          size={40}
+          testID={PerpsMarketHeaderSelectorsIDs.ASSET_ICON}
+        />
       }
-      title={useDetailLayout ? undefined : displayTitle}
-      titleProps={
-        useDetailLayout
-          ? undefined
-          : { testID: PerpsMarketHeaderSelectorsIDs.ASSET_NAME }
-      }
-      titleEndAccessory={useDetailLayout ? undefined : leverageBadge}
-      description={useDetailLayout ? undefined : compactDescription}
+      title={displayTitle}
+      titleProps={{ testID: PerpsMarketHeaderSelectorsIDs.ASSET_NAME }}
+      titleEndAccessory={leverageBadge}
+      description={compactDescription}
     />
   );
 };
