+++ app/components/UI/Perps/hooks/usePerpsProMarketHeaderActions.test.ts
@@ -10,6 +10,7 @@ import { usePerpsProMarketHeaderActions } from './usePerpsProMarketHeaderActions
 const mockNavigateBack = jest.fn();
 const mockNavigateToWallet = jest.fn();
 const mockNavigateToMarketList = jest.fn();
+const mockNavigateToMarketListFromHeader = jest.fn();
 let mockCanGoBack = true;
 
 jest.mock('./usePerpsNavigation', () => ({
@@ -17,7 +18,10 @@ jest.mock('./usePerpsNavigation', () => ({
     navigateBack: mockNavigateBack,
     navigateToWallet: mockNavigateToWallet,
     navigateToMarketList: mockNavigateToMarketList,
-    canGoBack: mockCanGoBack,
+    navigateToMarketListFromHeader: mockNavigateToMarketListFromHeader,
+    get canGoBack() {
+      return mockCanGoBack;
+    },
   })),
 }));
 
@@ -98,7 +102,7 @@ describe('usePerpsProMarketHeaderActions', () => {
       result.current.handleMarketListPress();
     });
 
-    expect(mockNavigateToMarketList).toHaveBeenCalledWith({
+    expect(mockNavigateToMarketListFromHeader).toHaveBeenCalledWith({
       source: PERPS_EVENT_VALUE.SOURCE.PERP_ASSET_SCREEN,
     });
     expect(mockTrack).toHaveBeenCalledWith(
@@ -120,7 +124,7 @@ describe('usePerpsProMarketHeaderActions', () => {
       result.current.handleMarketListPress();
     });
 
-    expect(mockNavigateToMarketList).not.toHaveBeenCalled();
+    expect(mockNavigateToMarketListFromHeader).not.toHaveBeenCalled();
     expect(mockTrack).not.toHaveBeenCalled();
   });
