+++ app/components/UI/Perps/types/navigation.ts
@@ -1,4 +1,5 @@
 import type { NavigatorScreenParams } from '@react-navigation/native';
+import type { NativeStackNavigationOptions } from '@react-navigation/native-stack';
 import {
   type Position,
   type Order,
@@ -138,6 +139,13 @@ export type PerpsStackParamList = {
         button_clicked?: string;
         button_location?: string;
         transactionActiveAbTests?: TransactionActiveAbTestEntry[];
+        animation?: NativeStackNavigationOptions['animation'];
+        /**
+         * When true, selecting a market replaces the underlying MARKET_DETAILS
+         * (and dismisses this list) instead of pushing another details screen.
+         * Used by the header slide-up picker.
+         */
+        replaceOnSelect?: boolean;
       }
     | undefined;
 
@@ -309,6 +317,7 @@ export type PerpsStackParamList = {
         button_clicked?: string;
         button_location?: string;
         transactionActiveAbTests?: TransactionActiveAbTestEntry[];
+        animation?: NativeStackNavigationOptions['animation'];
       }
     | undefined;
   PerpsOrderDetailsView: {
